(feat) #6 Addded tests to applyAddedDiffElement
Added a couple of tests to test continuity of lexical preservation when
adding an element after a comment and a new line which was not tested
before.

Also added a testcase where an element with a set range is added right
after a comment. This case does not seem to be handled well currently.