(feat) #14 Added another test
Added another test for the reduction between an inference variable and
another variable and an objecttype variable, the result should be a
"sameasbound" between the variables.

Also commented out the test for reduction with nulltypes. Turns out this
will always throw an error as it is right now because if the
isProperType method.